+++ esnextguardian.js
@@ -1,9 +1,9 @@
-// September 8, 2015
+// 9 September 2015
 // https://github.com/bevry/base
 if ( process.env.REQUIRE_ESNEXT ) {
 	module.exports = require('./esnext/lib/extendr.js')
 }
-else if ( !process.versions.v8 || process.versions.v8.split('.')[0] < 4 ) {
+else if ( process.env.REQUIRE_ES5 || !process.versions.v8 || process.versions.v8.split('.')[0] < 4 ) {
 	module.exports = require('./es5/lib/extendr.js')
 }
 else {